--- e2e/app_configs/remix_ts_app.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../helpers/test_runner_utils'
-
-class RemixTsApp < App
-  include TestRunnerUtils
-
-  def initialize(root_dir, *args)
-    super('remix_ts', root_dir, *args)
-  end
-
-  private
-
-  def yarn_create!
-    Dir.chdir(@root_dir) do
-      # install with the "remix" template
-      system_quiet('yarn create remix . --template=remix --typescript --install')
-
-      ## variation: indie-stack is a different remix template that people use
-      # system_quiet("yarn create remix . --template=remix-run/indie-stack --typescript --install")
-
-      # yarn add peer dependencies
-      system_quiet('yarn add next react react-dom')
-    end
-  end
-end
