+++ e2e/configs/jest_tests/package.json
